Fix a spawn-under-world glitch when using mixed teleportation methods (#28)
Fixes #27 

When a player exclusively leaves a dimension they reached via cake by some means other than a cake, their return position via cake is not cached. This resulted in the next use of that dimension's cake spawning the player at {0,0,0} and falling out of the world.

This patch sanity-checks the retrieved return position, recomputing the destination if the Y position would spawn the player under the world. Unless the destination has become obstructed, eating the same cake should place the player in the same position as before.